+++ src/builders/CacheBuilder.ts
@@ -1,8 +1,10 @@
 import App from "@bejibun/app";
 import Logger from "@bejibun/logger";
 import Redis from "@bejibun/redis";
+import Luxon from "@bejibun/utils/facades/Luxon";
 import {isEmpty, isNotEmpty} from "@bejibun/utils";
 import fs from "fs";
+import path from "path";
 import CacheConfig from "@/config/cache";
 import CacheException from "@/exceptions/CacheException";
 
@@ -29,13 +31,57 @@ export default class CacheBuilder {
     }
 
     private key(key: string): string {
-        return `${this.prefix}/${key}`;
+        const defaultKey: string = `${this.prefix}-${key.replaceAll("/", "-").replaceAll(" ", "-")}`;
+
+        return defaultKey;
+        /*if (forceDefault) return defaultKey;
+
+        switch (this.config.connection) {
+            case "local":
+                return `${Luxon.DateTime.now().toUnixInteger()}-${defaultKey}`;
+            default:
+                return defaultKey;
+        }*/
+    }
+
+    private connection(): any {
+        return this.config.connections[this.config.connection];
+    }
+
+    private filePath(key: string): string {
+        return path.resolve(this.connection().path, `${this.key(key)}.cache`);
+    }
+
+    private file(key: string): Bun.BunFile {
+        return Bun.file(this.filePath(key));
+    }
+
+    private async setFile(key: string, data: any): Promise<number> {
+        await fs.promises.mkdir(this.connection().path, {recursive: true});
+
+        return await Bun.write(this.filePath(key), data);
+    }
+
+    private async getFile(key: string): Promise<any> {
+        const file = this.file(key);
+
+        if (await file.exists()) return await file.text();
+
+        return null;
     }
 
     public async remember(key: string, callback: Function, ttl?: number): Promise<any> {
         let data: any;
 
         switch (this.config.connection) {
+            case "local":
+                data = await this.getFile(key);
+
+                if (isEmpty(data)) {
+                    data = callback();
+                    await this.setFile(key, data);
+                }
+                break;
             case "redis":
                 data = await Redis.get(this.key(key));
 
@@ -56,6 +102,9 @@ export default class CacheBuilder {
         let data: any;
 
         switch (this.config.connection) {
+            case "local":
+                data = await this.getFile(key);
+                break;
             case "redis":
                 data = await Redis.get(this.key(key));
                 break;
@@ -71,6 +120,9 @@ export default class CacheBuilder {
         let data: any;
 
         switch (this.config.connection) {
+            case "local":
+                data = await this.getFile(key);
+                break;
             case "redis":
                 data = await Redis.get(this.key(key));
                 break;
@@ -88,6 +140,9 @@ export default class CacheBuilder {
 
         try {
             switch (this.config.connection) {
+                case "local":
+                    data = await this.getFile(key);
+                    break;
                 case "redis":
                     data = await Redis.get(this.key(key));
                     break;
@@ -98,6 +153,9 @@ export default class CacheBuilder {
 
             if (isEmpty(data)) {
                 switch (this.config.connection) {
+                    case "local":
+                        await this.setFile(key, value);
+                        break;
                     case "redis":
                         await Redis.set(this.key(key), value, ttl);
                         break;
@@ -121,6 +179,9 @@ export default class CacheBuilder {
 
         try {
             switch (this.config.connection) {
+                case "local":
+                    await this.setFile(key, value);
+                    break;
                 case "redis":
                     await Redis.set(this.key(key), value, ttl);
                     break;
@@ -137,6 +198,13 @@ export default class CacheBuilder {
 
     public async forget(key: string): Promise<void> {
         switch (this.config.connection) {
+            case "local":
+                try {
+                    await this.file(key).delete();
+                } catch (error: any) {
+                    break;
+                }
+                break;
             case "redis":
                 await Redis.del(this.key(key));
                 break;
@@ -149,6 +217,17 @@ export default class CacheBuilder {
         let data: number;
 
         switch (this.config.connection) {
+            case "local":
+                data = Number(await this.getFile(key));
+
+                if (isEmpty(data)) {
+                    data = 1;
+                    await this.setFile(key, String(data));
+                } else {
+                    data++;
+                    await this.setFile(key, String(data));
+                }
+                break;
             case "redis":
                 data = Number(await Redis.get(this.key(key)));
 
@@ -172,6 +251,17 @@ export default class CacheBuilder {
         let data: number;
 
         switch (this.config.connection) {
+            case "local":
+                data = Number(await this.getFile(key));
+
+                if (isEmpty(data)) {
+                    data = -1;
+                    await this.setFile(key, String(data));
+                } else {
+                    data--;
+                    await this.setFile(key, String(data));
+                }
+                break;
             case "redis":
                 data = Number(await Redis.get(this.key(key)));