+++ src/config/cache.ts
@@ -1,7 +1,13 @@
+import App from "@bejibun/app";
+
 const config: Record<string, any> = {
-    connection: "redis",
+    connection: "local",
 
     connections: {
+        local: {
+            path: App.Path.storagePath("cache") // absolute path
+        },
+
         redis: {
             host: "127.0.0.100",
             port: 6379,