+++ CHANGELOG.md
@@ -3,6 +3,23 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [v0.1.12](https://github.com/crenata/bejibun-cache/compare/v0.1.11...v0.1.12) - 2025-12-04
+
+### 🩹 Fixes
+
+### 📖 Changes
+What's New :
+- Adding `local` connection for file schema
+
+Now, [@bejibun/cache](https://github.com/crenata/bejibun-cache) has local and redis for cache system.
+
+### ❤️Contributors
+- Havea Crenata ([@crenata](https://github.com/crenata))
+
+**Full Changelog**: https://github.com/crenata/bejibun-cache/blob/master/CHANGELOG.md
+
+---
+
 ## [v0.1.11](https://github.com/crenata/bejibun-cache/compare/v0.1.0...v0.1.11) - 2025-11-23
 
 ### 🩹 Fixes